requirements.txt
numpy==1.22.0
# tensorflow==2.8
tensorflow
pandas
scipy
pillow
matplotlib
requests
geocoder
torch